Newer code for measuring performance impact of metrics
